Ft: Build respon handler for Ok & Error

+++ Controllers/UniversityController.cs
@@ -1,7 +1,10 @@
 ﻿using Booking_Api.Contracts;
+using Booking_Api.DTOs.Employees;
 using Booking_Api.DTOs.Universities;
 using Booking_Api.Models;
+using Booking_Api.Utilities.Handler;
 using Microsoft.AspNetCore.Mvc;
+using System.Net;
 
 namespace Booking_Api.Controllers
 {
@@ -22,12 +25,17 @@ namespace Booking_Api.Controllers
             var universities = _universityRepository.GetAll();
             if (!universities.Any())
             {
-                return NotFound("Data Not Found");
+                return NotFound(new ResponseErrorHandler
+                {
+                    Code = StatusCodes.Status404NotFound,
+                    Status = HttpStatusCode.NotFound.ToString(),
+                    Message = "Data Not Found"
+                });
             }
 
             var data = universities.Select(x => (UniversityDto)x);
 
-            return Ok(data);
+            return Ok(new ResponseOKHandler<IEnumerable<UniversityDto>>(data));
         }
 
         [HttpGet("{guid}")]
@@ -36,54 +44,99 @@ namespace Booking_Api.Controllers
             var university = _universityRepository.GetById(guid);
             if (university is null)
             {
-                return NotFound("University Not found");
+                return NotFound(new ResponseErrorHandler
+                {
+                    Code = StatusCodes.Status404NotFound,
+                    Status = HttpStatusCode.NotFound.ToString(),
+                    Message = "Data Not Found"
+                });
             }
-            return Ok((UniversityDto)university);
+            return Ok(new ResponseOKHandler<UniversityDto>((UniversityDto)university));
         }
 
         [HttpPost]
         public IActionResult Create(CreateUniversityDto createUniversityDto)
         {
+            try { 
             var createdUniversity = _universityRepository.Create(createUniversityDto);
-            if (createdUniversity is null)
-            {
-                return BadRequest("Not Created University. Try Again!");
+            
+            return Ok(new ResponseOKHandler<UniversityDto>((UniversityDto)createdUniversity));
             }
-            return Ok((UniversityDto)createdUniversity);
+            catch (ExceptionHandler ex)
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError, new ResponseErrorHandler
+                {
+                    Code = StatusCodes.Status500InternalServerError,
+                    Status = HttpStatusCode.InternalServerError.ToString(),
+                    Message = "Failed to create data",
+                    Error = ex.Message
+                });
+            }
         }
 
         [HttpPut]
         public IActionResult Update(UniversityDto universityDto)
         {
+            try { 
             var university = _universityRepository.GetById(universityDto.Guid);
             if (university is null)
             {
-                return NotFound("Id Not Found");
-            }
+                    return NotFound(new ResponseErrorHandler
+                    {
+                        Code = StatusCodes.Status404NotFound,
+                        Status = HttpStatusCode.NotFound.ToString(),
+                        Message = "Data Not Found"
+                    });
+                }
 
             University toUpdate = universityDto;
             toUpdate.CreatedDate = university.CreatedDate;
 
-            var updatedUniversity = _universityRepository.Update(university);
-            if (!updatedUniversity)
-            {
-                return BadRequest("Not Updated University. Try Again!");
+            _universityRepository.Update(university);
+            
+            return Ok(new ResponseOKHandler<string>("Data has been updated successfully"));
             }
-            return Ok("Data has been updated successfully");
+            catch (ExceptionHandler ex) {
+                return StatusCode(StatusCodes.Status500InternalServerError, new ResponseErrorHandler
+                {
+                    Code = StatusCodes.Status500InternalServerError,
+                    Status = HttpStatusCode.InternalServerError.ToString(),
+                    Message = "Failed to create data",
+                    Error = ex.Message
+                });
+            }
         }
 
         [HttpDelete("{guid}")]
         public IActionResult Delete(Guid guid)
         {
-           
-            var university = _universityRepository.GetById(guid);
-
-            var deletedUniversity = _universityRepository.Delete(university);
-            if (!deletedUniversity)
+            try
             {
-                return BadRequest("Not Deleted University. Try Again!");
+                var university = _universityRepository.GetById(guid);
+                if (university is null)
+                {
+                    return NotFound(new ResponseErrorHandler
+                    {
+                        Code = StatusCodes.Status404NotFound,
+                        Status = HttpStatusCode.NotFound.ToString(),
+                        Message = "Data Not Found"
+                    });
+                }
+
+                _universityRepository.Delete(university);
+          
+                return Ok(new ResponseOKHandler<string>("Data has been Deleted successfully"));
             }
-            return Ok(deletedUniversity);
+            catch (ExceptionHandler ex)
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError, new ResponseErrorHandler
+                {
+                    Code = StatusCodes.Status500InternalServerError,
+                    Status = HttpStatusCode.InternalServerError.ToString(),
+                    Message = "Failed to create data",
+                    Error = ex.Message
+                });
+            }
         }
     }
 }
